add :Output section to resource declarations
Previously, templates could get unnecessarily verbose when defining a
resource and its associated outputs. eg:

    resource "AResource", {:Properties...}
    output "AResourceId", {:Value => ref("AResource")
    output "AResourceProperty", {:Value => get_att("AResource", ...)}

This is not only tedious to read and write, the verbosity of the whole
process leads to copy&pasting, which leads to copy&paste errors.

To correct this, override the resource function, to process a special
:Output key of the options hash.

Basic Usage: instead of,

    resource "MyEIP", :Properties...
    output "MyEIPAllocationId",
        :Value => ref("MyEIP"),
        :Description => "The AllocationId of My EIP"

do:

    resource "MyEIP", :Properties..., :Output => [:AllocationId]

The default formatting / phrasing is based on those used internally by
the organisation for which this helper function was created, but these
are expected to be generically useful.