add :Export key for resources
Previously, an :Output key was added for resources. This would allow for
the quick definition of outputs to go along with resources.

However, almost every resource would be desired as an Export, as well as
an Output, meaning the Output definitions would almost always read
something like:

:Output => {:Export => true, :Attributes => {:Ref => "Arn"}}

To handle this common case, add an :Export key, as a shorthand for the
above. So, the above can now be written as:

:Export => {:Ref => "Arn"}