only output parameters when we have something to output
previously, we output parameters only when the paramters themselves
said they they needed to be (ie: having no default, or NoEcho).

Then, in 316e780d4a2ab161d505f0cb106c149edf419dc4 we switched this
around: *any* parameter would be output, potentially with a value which
we defined.

This was a step to far: This had us outputting parameters even when we
had no value available within them. In theory `self.parameters` should
have had something - but digging through the code for
cloudformation-ruby-dsl reveals that this value defaults to the empty
string - which cloudformation-ruby-dsl specifically checks in order to
decide if the value should be output ("value.empty?")

This smells bad, and so was not previously included - after all, we
could easily want to override a default value with an empty value at
some point.

That said, in production this is very unlikely to come up, and
compatibility with cloudformation-ruby-dsl will bring a bit more sanity.

More to the point, it will restore out ability to *not* output all
parameters, allowing defaults to be used again.