fix handling of nil template_shell
when the shell-specific Nickname template was empty, we were meant to
use the "default" Nickname template instead. However, when the
shell-specific Nickname template was empty, we first tried to perform
the backslash-escaping. This: 1) failed on nil values, and 2) would have
resulted in a non-nil result, in any case.

Instead, immediately switch to using the "default" Nickname template if
the shell-specific template is nil.