teach ECS partial about overlay2 driver
ECS Optimised instances default to using the "devicemapper" storage
driver for docker. This default is useful in a large cluster in which
the images are not known in advance, but is extremely inefficient in
situations where most containers are known to share the same image.

Using the overlay2 driver allows these "shared" images to be shared in
memory as well, vastly increasing memory efficiency and vastly reducing
disk IO requirements. In theory. This change is being written based on
the docker storage driver documentation, and has not yet been tested in
a "real-world" situation.